requirements.txt
fastapi
uvicorn
pydantic
yahooquery
langchain
openai
python-dotenv